requirements.txt
requests==2.25.1
tweepy==3.10.0
PyGithub==1.55
PyYAML==5.4.1
